问题回复:异常处理、事务、页面提交回复

+++ src/main/java/com/example/moon/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
+    public String edit(@PathVariable(name="id") Long id,
                        Model model,
                        HttpServletRequest request){
         QuestionDTO question = questionService.getById(id);
@@ -49,7 +49,7 @@ public class PublishController {
             @RequestParam("title")String title,
             @RequestParam("description")String description,
             @RequestParam("tag")String tag,
-            @RequestParam("id")Integer id,
+            @RequestParam("id")Long id,
             HttpServletRequest request,
             Model model){
